Preparation for more 64-bit versions

+++ SR/llasm-support/llasm_float.c
@@ -374,7 +374,7 @@ EXTERNC int32_t x87_fist_int32(CPU)
             }
             else
             {
-                ival = 0x80000000;
+                ival = (int32_t) 0x80000000;
             }
             return ival;
         }
@@ -390,7 +390,7 @@ EXTERNC int32_t x87_fist_int32(CPU)
         break;
     }
     CLEAR_X87_FLAGS;
-    return ((dval < 2147483648.0) && (dval > -2147483648.0))?((int32_t) dval):0x80000000;
+    return ((dval < 2147483648.0) && (dval > -2147483648.0))?((int32_t) dval):((int32_t) 0x80000000);
 }
 
 EXTERNC int16_t x87_fistp_int16(CPU)
@@ -417,7 +417,7 @@ EXTERNC int16_t x87_fistp_int16(CPU)
             }
             else
             {
-                ival = 0x8000;
+                ival = (int16_t) 0x8000;
             }
             return ival;
         }
@@ -433,7 +433,7 @@ EXTERNC int16_t x87_fistp_int16(CPU)
         break;
     }
     POP_REGS;
-    return ((dval < 32768.0) && (dval > -32768.0))?((int16_t) dval):0x8000;
+    return ((dval < 32768.0) && (dval > -32768.0))?((int16_t) dval):((int16_t) 0x8000);
 }
 
 EXTERNC int32_t x87_fistp_int32(CPU)
@@ -460,7 +460,7 @@ EXTERNC int32_t x87_fistp_int32(CPU)
             }
             else
             {
-                ival = 0x80000000;
+                ival = (int32_t) 0x80000000;
             }
             return ival;
         }
@@ -476,7 +476,7 @@ EXTERNC int32_t x87_fistp_int32(CPU)
         break;
     }
     POP_REGS;
-    return ((dval < 2147483648.0) && (dval > -2147483648.0))?((int32_t) dval):0x80000000;
+    return ((dval < 2147483648.0) && (dval > -2147483648.0))?((int32_t) dval):((int32_t) 0x80000000);
 }
 
 EXTERNC uint32_t x87_fistp_int64(CPU)
@@ -1007,7 +1007,7 @@ EXTERNC int32_t x87_ftol_int32(CPU)
 
     if (num1 < 0)
     {
-        if (num1 <= -2147483648.0) return 0x80000000;
+        if (num1 <= -2147483648.0) return (int32_t) 0x80000000;
 
         result.d = num1 + doublemagic;  // fast conversion to int,
         num2 = (double)(int32_t)result.low; // result.low contains the result (rounded up or down)
@@ -1021,10 +1021,10 @@ EXTERNC int32_t x87_ftol_int32(CPU)
     }
     else
     {
-        if (num1 >= 2147483648.0) return 0x80000000;
+        if (num1 >= 2147483648.0) return (int32_t) 0x80000000;
 
         result.d = num1 + doublemagic;  // fast conversion to int,
-        if (0 > (int32_t)result.low) return 0x7fffffff;
+        if (0 > (int32_t)result.low) return (int32_t) 0x7fffffff;
         num2 = (double)(int32_t)result.low; // result.low contains the result (rounded up or down)
 
         if (num2 > num1) // compare result with original value and if the result was rounded toward positive infinity, then decrease result (truncate toward 0)
