+++ games/Warcraft/SR-War/xmi2mid.c
@@ -1,6 +1,6 @@
 /**
  *
- *  Copyright (C) 2016 Roman Pauer
+ *  Copyright (C) 2016-2026 Roman Pauer
  *
  *  Permission is hereby granted, free of charge, to any person obtaining a copy of
  *  this software and associated documentation files (the "Software"), to deal in
@@ -384,7 +384,7 @@ uint8_t *xmi2mid(uint8_t *xmi, int seq_num, unsigned int *midi_size)
                 xmi+=interval;
 
                 cur_event->time = last_time;
-                cur_event->length = xmi - event_data;
+                cur_event->length = (int32_t)(xmi - event_data);
                 cur_event->data = event_data;
                 cur_event++;
                 num_events++;
